Split DRM internal interface to its own header
