+++ backend/drm/iface_legacy.c
@@ -3,6 +3,7 @@
 #include <xf86drmMode.h>
 #include <wlr/util/log.h>
 #include "backend/drm/drm.h"
+#include "backend/drm/iface.h"
 #include "backend/drm/util.h"
 
 static bool legacy_crtc_pageflip(struct wlr_drm_backend *backend,
