+++ include/backend/drm/drm.h
@@ -15,6 +15,7 @@
 #include <wlr/egl.h>
 #include <wlr/util/list.h>
 
+#include "iface.h"
 #include "properties.h"
 
 struct wlr_drm_plane {
@@ -154,23 +155,6 @@ struct wlr_drm_output {
 
 	bool pageflip_pending;
 	struct wl_event_source *retry_pageflip;
-};
-
-// Used to provide atomic or legacy DRM functions
-struct wlr_drm_interface {
-	// Enable or disable DPMS for output
-	void (*conn_enable)(struct wlr_drm_backend *backend,
-			struct wlr_drm_output *output, bool enable);
-	// Pageflip on crtc. If mode is non-NULL perform a full modeset using it.
-	bool (*crtc_pageflip)(struct wlr_drm_backend *backend,
-			struct wlr_drm_output *output, struct wlr_drm_crtc *crtc,
-			uint32_t fb_id, drmModeModeInfo *mode);
-	// Enable the cursor buffer on crtc. Set bo to NULL to disable
-	bool (*crtc_set_cursor)(struct wlr_drm_backend *backend,
-			struct wlr_drm_crtc *crtc, struct gbm_bo *bo);
-	// Move the cursor on crtc
-	bool (*crtc_move_cursor)(struct wlr_drm_backend *backend,
-			struct wlr_drm_crtc *crtc, int x, int y);
 };
 
 bool wlr_drm_check_features(struct wlr_drm_backend *drm);
